4.15 - 4.23, all tests passed

+++ part4/blogilista/backend/controllers/blogs.js
@@ -1,16 +1,21 @@
 const blogsRouter = require('express').Router()
 const { response, request } = require('../app')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
   response.json(blogs)
   
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
+
+  const user = request.user
 
   if(!body.title){
     return response.status(400).json({
@@ -24,21 +29,45 @@ blogsRouter.post('/', async (request, response) => {
     })
   }
 
+  
+  if (!user) {
+    return response.status(400).json({ error: 'userId missing or not valid' })
+  }
+
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: user._id 
   })
 
   const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
   
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === user.id.toString()){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else{
+    return response.status(400).json({ error: 'cannot remove blogs you have not made' })
+  }
+  
 })
 
 blogsRouter.put('/:id', async(request, response) =>{
